moved pages to be outside of the components folder
