+++ client/src/Pages/AdminArea/AdminArea.js
@@ -1,8 +1,8 @@
 import React from "react";
 
 // import Nav from "../AdminNav/AdminNav";
-import CardHomeInfo from "../../CardHomeInfo/CardHomeInfo";
-import AdminNav from "./AdminNav/AdminNav";
+//import CardHomeInfo from "../../components/CardHomeInfo/CardHomeInfo";
+import AdminNav from "../../components/AdminNav/AdminNav";
 
 const AdminArea = () => {
     return (
