+++ client/src/Pages/MainPage/Main.js
@@ -1,10 +1,10 @@
 import React, { Component } from "react";
-import API from "../../../utils/API"
-import SearchForm from "../../searchform/SearchForm.js";
+import API from "../../utils/API"
+import SearchForm from "../../components/searchform/SearchForm.js";
 import {MDBRow, MDBCol, MDBContainer } from 'mdbreact';
-import Nav from "../../Nav/Nav";
-import CardHomeInfo from "../../CardHomeInfo/CardHomeInfo";
-import MapContainer from "../../MapContainer/MapContainer"
+import Nav from "../../components/Nav/Nav";
+import CardHomeInfo from "../../components/CardHomeInfo/CardHomeInfo";
+import MapContainer from "../../components/MapContainer/MapContainer"
 
 
 class MainPage extends Component {
