+++ client/src/components/AdminNav/AdminNav.js
@@ -1,10 +1,10 @@
 import React, { Component } from "react";
-import { BrowserRouter as Router, NavLink} from 'react-router-dom';
+import { NavLink} from 'react-router-dom';
 import { MDBNavbar, MDBNavbarBrand, MDBNavbarNav, MDBNavItem, MDBNavLink, MDBNavbarToggler, MDBCollapse, MDBDropdown,
 MDBDropdownToggle, MDBDropdownMenu, MDBDropdownItem, MDBIcon } from "mdbreact";
 
 
-import "../AdminNav/AdminNav.css";
+import "./AdminNav.css";
 
 class AdminNav extends Component {
 state = {
